refactor to use more modern c++
Parse Json and CSV data without knowing the fields name.
For now expect json dat to be in a certain format. The field name and
data are in a string format. will work on parsing
all json type specially where the type is in a metadata and athe fiels
only contain the data